+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.5.0
-- Added Division Preview button and divPreviews.js
+- Added Division Previews
 - updated logoHandler to imageHandler
+- New modular formula for properly spacing variable team blocks vertically in space (divPreviews.js)
 
 ## 2.4.1
 - Folder structure: Gameday Graphics is now expected directly inside the league folder (no Automations folder)